+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 3.0.0-beta.1
+* Added support for a Kidding Schedule!
+  * Includes a built-in gestation calculator (currently uses 145 days for Nigerian Dwarfs & 150 days for all other breeds)
+  * Displays the age of your doe when was bred & is due
+
 ## 2.1.14-beta.1
 * Fixed the version reported when installing web-ui updates